+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/balances/view/BalancesView.kt
@@ -1,25 +1,40 @@
+@file:Suppress("TooManyFunctions")
+
 package co.electriccoin.zcash.ui.screen.balances.view
 
+import androidx.compose.animation.animateContentSize
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Divider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
@@ -32,6 +47,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import cash.z.ecc.android.sdk.Synchronizer
 import cash.z.ecc.android.sdk.model.FiatCurrencyConversionRateState
+import cash.z.ecc.android.sdk.model.Zatoshi
 import cash.z.ecc.android.sdk.model.toZecString
 import cash.z.ecc.sdk.extension.toPercentageWithDecimal
 import co.electriccoin.zcash.ui.R
@@ -49,11 +65,14 @@ import co.electriccoin.zcash.ui.design.component.CircularScreenProgressIndicator
 import co.electriccoin.zcash.ui.design.component.CircularSmallProgressIndicator
 import co.electriccoin.zcash.ui.design.component.GradientSurface
 import co.electriccoin.zcash.ui.design.component.LinearProgressIndicator
+import co.electriccoin.zcash.ui.design.component.PrimaryButton
+import co.electriccoin.zcash.ui.design.component.Reference
 import co.electriccoin.zcash.ui.design.component.SmallTopAppBar
 import co.electriccoin.zcash.ui.design.component.StyledBalance
 import co.electriccoin.zcash.ui.design.theme.ZcashTheme
 import co.electriccoin.zcash.ui.fixture.WalletSnapshotFixture
 import co.electriccoin.zcash.ui.screen.balances.BalancesTag
+import co.electriccoin.zcash.ui.screen.balances.model.ShieldState
 import co.electriccoin.zcash.ui.screen.balances.model.WalletDisplayValues
 
 @Preview("Balances")
@@ -66,18 +85,23 @@ private fun ComposablePreview() {
                 isFiatConversionEnabled = false,
                 isKeepScreenOnWhileSyncing = false,
                 isUpdateAvailable = false,
+                onShielding = {},
                 walletSnapshot = WalletSnapshotFixture.new(),
+                shieldState = ShieldState.Available,
             )
         }
     }
 }
 
+@Suppress("LongParameterList")
 @Composable
 fun Balances(
     onSettings: () -> Unit,
     isFiatConversionEnabled: Boolean,
     isKeepScreenOnWhileSyncing: Boolean?,
     isUpdateAvailable: Boolean,
+    onShielding: () -> Unit,
+    shieldState: ShieldState,
     walletSnapshot: WalletSnapshot?,
 ) {
     Scaffold(topBar = {
@@ -90,14 +114,16 @@ fun Balances(
                 isFiatConversionEnabled = isFiatConversionEnabled,
                 isKeepScreenOnWhileSyncing = isKeepScreenOnWhileSyncing,
                 isUpdateAvailable = isUpdateAvailable,
+                onShielding = onShielding,
                 walletSnapshot = walletSnapshot,
+                shieldState = shieldState,
                 modifier =
                     Modifier.padding(
                         top = paddingValues.calculateTopPadding() + ZcashTheme.dimens.spacingDefault,
                         bottom = paddingValues.calculateBottomPadding() + ZcashTheme.dimens.spacingHuge,
                         start = ZcashTheme.dimens.screenHorizontalSpacingRegular,
                         end = ZcashTheme.dimens.screenHorizontalSpacingRegular
-                    )
+                    ),
             )
         }
     }
@@ -122,12 +148,15 @@ private fun BalancesTopAppBar(onSettings: () -> Unit) {
     )
 }
 
+@Suppress("LongParameterList")
 @Composable
 private fun BalancesMainContent(
     isFiatConversionEnabled: Boolean,
     isKeepScreenOnWhileSyncing: Boolean?,
     isUpdateAvailable: Boolean,
+    onShielding: () -> Unit,
     walletSnapshot: WalletSnapshot,
+    shieldState: ShieldState,
     modifier: Modifier = Modifier,
 ) {
     Column(
@@ -162,20 +191,11 @@ private fun BalancesMainContent(
 
         Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingLarge))
 
-        Column(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .height(166.dp)
-                    .background(color = ZcashTheme.colors.panelBackgroundColor),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Body(
-                text = stringResource(id = R.string.balances_coming_soon),
-                textAlign = TextAlign.Center
-            )
-        }
+        TransparentBalancePanel(
+            onShielding = onShielding,
+            shieldState = shieldState,
+            walletSnapshot = walletSnapshot,
+        )
 
         Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingDefault))
 
@@ -189,6 +209,162 @@ private fun BalancesMainContent(
         if (isKeepScreenOnWhileSyncing == true && walletSnapshot.status == Synchronizer.Status.SYNCING) {
             DisableScreenTimeout()
         }
+    }
+}
+
+@Composable
+fun TransparentBalancePanel(
+    onShielding: () -> Unit,
+    shieldState: ShieldState,
+    walletSnapshot: WalletSnapshot,
+) {
+    var showHelpPanel by rememberSaveable { mutableStateOf(false) }
+
+    // TODO [#1242]: Create error popup UI
+    // TODO [#1242]: https://github.com/Electric-Coin-Company/zashi-android/issues/1242
+    // if (shieldState is ShieldState.Failed) {
+    // }
+
+    Box(
+        modifier =
+            Modifier
+                .background(color = ZcashTheme.colors.panelBackgroundColor)
+                .wrapContentSize()
+                .animateContentSize()
+    ) {
+        Column(horizontalAlignment = Alignment.CenterHorizontally) {
+            Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingDefault))
+
+            TransparentBalanceRow(
+                isProgressbarVisible = shieldState == ShieldState.Running,
+                onHelpClick = { showHelpPanel = !showHelpPanel },
+                walletSnapshot = walletSnapshot
+            )
+
+            Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingDefault))
+
+            PrimaryButton(
+                onClick = onShielding,
+                text = stringResource(R.string.balances_transparent_balance_shield),
+                textStyle = ZcashTheme.extendedTypography.buttonTextSmall,
+                enabled = shieldState == ShieldState.Available,
+                outerPaddingValues =
+                    PaddingValues(
+                        horizontal = 54.dp,
+                        vertical = ZcashTheme.dimens.spacingSmall
+                    )
+            )
+
+            Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingDefault))
+
+            BodySmall(
+                text =
+                    stringResource(
+                        id = R.string.balances_transparent_balance_fee,
+                        // TODO [#1047]: Representing Zatoshi amount
+                        // TODO [#1047]: https://github.com/Electric-Coin-Company/zashi-android/issues/1047
+                        @Suppress("MagicNumber")
+                        Zatoshi(100_000L).toZecString()
+                    ),
+                textFontWeight = FontWeight.SemiBold
+            )
+
+            Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingXlarge))
+        }
+
+        if (showHelpPanel) {
+            TransparentBalanceHelpPanel(
+                onHideHelpPanel = { showHelpPanel = !showHelpPanel }
+            )
+        }
+    }
+}
+
+@Composable
+fun TransparentBalanceRow(
+    isProgressbarVisible: Boolean,
+    onHelpClick: () -> Unit,
+    walletSnapshot: WalletSnapshot,
+) {
+    Row(
+        modifier =
+            Modifier
+                .fillMaxWidth()
+                .padding(
+                    start = ZcashTheme.dimens.spacingDefault,
+                    end = ZcashTheme.dimens.spacingSmall
+                ),
+        horizontalArrangement = Arrangement.SpaceBetween,
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        // To keep both elements together in relatively sized row
+        Row(modifier = Modifier.fillMaxWidth(TEXT_PART_WIDTH_RATIO)) {
+            // Apply common click listener
+            Row(
+                modifier =
+                    Modifier
+                        .clip(RoundedCornerShape(ZcashTheme.dimens.smallRippleEffectCorner))
+                        .clickable { onHelpClick() }
+            ) {
+                BodySmall(text = stringResource(id = R.string.balances_transparent_balance).uppercase())
+
+                Image(
+                    imageVector = ImageVector.vectorResource(R.drawable.ic_help_question_mark),
+                    contentDescription = stringResource(id = R.string.balances_transparent_help_content_description),
+                    modifier = Modifier.padding(horizontal = ZcashTheme.dimens.spacingXtiny)
+                )
+            }
+        }
+
+        Row(verticalAlignment = Alignment.CenterVertically) {
+            StyledBalance(
+                balanceString = walletSnapshot.transparentBalance.toZecString(),
+                textStyles =
+                    Pair(
+                        ZcashTheme.extendedTypography.balanceSingleStyles.first,
+                        ZcashTheme.extendedTypography.balanceSingleStyles.second
+                    ),
+                textColor = ZcashTheme.colors.textPending
+            )
+
+            Spacer(modifier = Modifier.width(ZcashTheme.dimens.spacingTiny))
+
+            Box(Modifier.width(ZcashTheme.dimens.circularSmallProgressWidth)) {
+                if (isProgressbarVisible) {
+                    CircularSmallProgressIndicator()
+                }
+            }
+        }
+    }
+}
+
+@Composable
+fun TransparentBalanceHelpPanel(onHideHelpPanel: () -> Unit) {
+    Column(
+        modifier =
+            Modifier
+                .padding(all = ZcashTheme.dimens.spacingDefault)
+                .background(color = Color.White)
+                .fillMaxSize(),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingDefault))
+
+        BodySmall(
+            text = stringResource(id = R.string.balances_transparent_balance_help),
+            textAlign = TextAlign.Center,
+            modifier = Modifier.padding(horizontal = ZcashTheme.dimens.spacingDefault)
+        )
+
+        Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingDefault))
+
+        Reference(
+            text = stringResource(id = R.string.balances_transparent_balance_help_close).uppercase(),
+            onClick = onHideHelpPanel,
+            textStyle = ZcashTheme.extendedTypography.referenceSmall
+        )
+
+        Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingDefault))
     }
 }