+++ ui-screenshot-test/src/main/java/co/electroniccoin/zcash/ui/screenshot/ScreenshotTest.kt
@@ -398,7 +398,7 @@ private fun balancesScreenshots(
     // TODO [#1127]: Implement Balances screen
     // TODO [#1127]: https://github.com/Electric-Coin-Company/zashi-android/issues/1127
 
-    composeTestRule.onNodeWithText(resContext.getString(R.string.balances_coming_soon)).also {
+    composeTestRule.onNodeWithText(resContext.getString(R.string.balances_title)).also {
         it.assertExists()
         ScreenshotTest.takeScreenshot(tag, "Balances 1")
     }