+++ CHANGELOG.md
@@ -14,8 +14,9 @@ directly impact users rather than highlighting other key architectural updates.*
 
 ### Added
 - The Balances screen now provides details on current balances like Change pending and Pending transactions
-- The screen also adds a new Block synchronization progress bar and status, which were initially part of the Account 
-  screen and redesigned
+- The Balances screen adds a new Block synchronization progress bar and status, which were initially part of the 
+  Account screen and redesigned
+- The Balances screen supports transparent funds shielding within its new shielding panel
 
 ### Fixed
 - Fixed character replacement in Zcash addresses on the Receive screen caused by ligatures in the app's primary font